\name{snp_data}
\alias{snp_data}
\docType{data}
\title{
Example SNP Data Set
}
\description{
An example snp data set.
}
\usage{data("snp_data")}
\format{
  A data frame with 1707 observations on the following 25 variables.
  \describe{
    \item{\samp{26463074}}{a numeric vector}
    \item{\samp{26463561}}{a numeric vector}
    \item{\samp{26464530}}{a numeric vector}
    \item{\samp{26464828}}{a numeric vector}
    \item{\samp{26465583}}{a numeric vector}
    \item{\samp{26465774}}{a numeric vector}
    \item{\samp{26467796}}{a numeric vector}
    \item{\samp{26468200}}{a numeric vector}
    \item{\samp{26469908}}{a numeric vector}
    \item{\samp{26474599}}{a numeric vector}
    \item{\samp{26475212}}{a numeric vector}
    \item{\samp{26475408}}{a numeric vector}
    \item{\samp{26475524}}{a numeric vector}
    \item{\samp{26475525}}{a numeric vector}
    \item{\samp{26475914}}{a numeric vector}
    \item{\samp{26476184}}{a numeric vector}
    \item{\samp{26476439}}{a numeric vector}
    \item{\samp{26477750}}{a numeric vector}
    \item{\samp{26478445}}{a numeric vector}
    \item{\samp{26480355}}{a numeric vector}
    \item{\samp{26480734}}{a numeric vector}
    \item{\samp{26480891}}{a numeric vector}
    \item{\samp{26480910}}{a numeric vector}
    \item{\samp{26462642}}{a numeric vector}
    \item{\samp{26463438}}{a numeric vector}
  }
}
\examples{
data(snp_data)
}
\keyword{datasets}
